Add defaults to options and exit the script gracefully with user-friendly messages.
* CLI options now have defaults, except `--host` and `--version`.
* Default values for CLI options are now shown in `--help`.
* Replaced all default Python exceptions with exceptions provided by the `click`
  package. This removes the unnecessary stack traces, when triggering an error.

+++ benchmark/generate.py
@@ -1,17 +1,16 @@
+import os
 from shutil import copyfile
 
 import click
 import datreant.core as dtr
 import mdsynthesis as mds
+from jinja2.exceptions import TemplateNotFound
 
 from .cli import cli
 from .util import ENV, get_possible_hosts, normalize_host
 
 
 def write_bench(top, tmpl, nodes, gpu, version, name, host, time):
-    if time > 1440:
-        raise ValueError('Cannot create batch jobs running longer than 24h.')
-
     sim = mds.Sim(
         top['{}/'.format(nodes)],
         categories={
@@ -23,10 +22,19 @@ def write_bench(top, tmpl, nodes, gpu, version, name, host, time):
             'name': name
         })
     sim.universedef.topology = name + '.tpr'
-    sim.universedef.trajectory = name + '.xtc'
+    # sim.universedef.trajectory = name + '.xtc'
+
     # copy input files
     mdp = '{}.mdp'.format(name)
     tpr = '{}.tpr'.format(name)
+
+    if not os.path.exists(mdp):
+        raise click.FileError(
+            mdp, hint='File does not exist or is not readable.')
+    if not os.path.exists(tpr):
+        raise click.FileError(
+            tpr, hint='File does not exist or is not readable.')
+
     copyfile(tpr, sim[tpr].relpath)
     copyfile(mdp, sim[mdp].relpath)
 
@@ -48,17 +56,30 @@ def write_bench(top, tmpl, nodes, gpu, version, name, host, time):
 
 
 @cli.command()
-@click.option('-n', '--name', help='name of tpr/mdp file')
-@click.option('-g', '--gpu', is_flag=True, help='run on gpu as well')
+@click.option(
+    '-n',
+    '--name',
+    help='name of tpr/mdp file',
+    default='md',
+    show_default=True)
+@click.option(
+    '-g', '--gpu', is_flag=True, help='run on gpu as well', show_default=True)
 @click.option('-v', '--version', help='gromacs module to use', multiple=True)
 @click.option('-h', '--host', help='job template name', default=None)
-@click.option('-m', '--max-nodes', help='test up to n nodes', type=int)
+@click.option(
+    '-m',
+    '--max-nodes',
+    help='test up to n nodes',
+    default=5,
+    show_default=True,
+    type=int)
 @click.option(
     '-t',
     '--time',
     help='run time for benchmark in minutes',
-    type=int,
-    default=15)
+    default=15,
+    show_default=True,
+    type=click.IntRange(1, 1440))
 @click.option('-l', '--list-hosts', help='show known hosts', is_flag=True)
 def generate(name, gpu, version, host, max_nodes, time, list_hosts):
     if list_hosts:
@@ -66,7 +87,17 @@ def generate(name, gpu, version, host, max_nodes, time, list_hosts):
         return
 
     host = normalize_host(host)
-    tmpl = ENV.get_template(host)
+    try:
+        tmpl = ENV.get_template(host)
+    except TemplateNotFound:
+        raise click.BadParameter(
+            'Could not find template for host \'{}\'.'.format(host),
+            param_hint='"-h" / "--host"')
+
+    if not version:
+        raise click.BadParameter(
+            'You did not specify which gromacs version to use for scaling.',
+            param_hint='"-v" / "--version"')
 
     for v in version:
         top_folder = '{}_{}'.format(host, v)
